+++ pkg/telemetry/trace.go
@@ -1,0 +1,23 @@
+package telemetry
+
+import (
+	"context"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
+	"go.opentelemetry.io/otel/sdk/resource"
+	"go.opentelemetry.io/otel/sdk/trace"
+)
+
+// creates a tracer provider with otlpgrpc exporter
+// TODO: Add support for multiple exporters
+func newTracerProvider(ctx context.Context, res *resource.Resource) (*trace.TracerProvider, error) {
+	traceExporter, err := otlptracegrpc.New(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	tracerProvider := trace.NewTracerProvider(
+		trace.WithResource(res),
+		trace.WithBatcher(traceExporter),
+	)
+	return tracerProvider, nil
+}
